# feat(perf): Improves the initial time to load the page