--- a/gatsby-config.js
+++ b/gatsby-config.js
@@ -61,6 +61,19 @@ const plugins = [
 			theme_color: `#343f44`,
 			display: `minimal-ui`,
 			icon: `static/favicon.png`,
+			icons: [
+				{
+				  src: `static/favicon.pngg`,
+				  sizes: `192x192`,
+				  type: `image/png`,
+				  purpose: "any maskable"
+				},
+				{
+				  src: `static/favicon.png`,
+				  sizes: `512x512`,
+				  type: `image/png`,
+				},
+			],
 		},
 	},
 	`gatsby-plugin-offline`,