--- a/gatsby-node.js
+++ b/gatsby-node.js
@@ -9,3 +9,32 @@ exports.createPages = ({ graphql, actions }) =>
 
 exports.onCreateNode = async ({ node, getNode, actions }) =>
 	await onCreateNode.createMarkdownNodeFields({ node, getNode, actions });
+
+
+exports .onCreateWebpackConfig = ({
+	stage,
+	rules,
+	loaders,
+	plugins,
+	actions,
+  }) => {
+	  actions.setWebpackConfig({
+		optimization: {
+			splitChunks: {
+				name: false,
+				chunks: `all`,
+				cacheGroups: {
+					default: false,
+					vendors: false,
+					commons: {
+					  name: `commons`,
+					  chunks: `all`,
+					  // if a chunk is used more than half the components count,
+					  // we can assume it's pretty global
+					  minChunks: 2,
+					},
+				}
+			}
+		}
+	  })
+  }