--- a/package.json
+++ b/package.json
@@ -66,6 +66,7 @@
 		"@appbaseio/designkit": "^0.9.12",
 		"@appbaseio/reactivemaps": "^3.0.0-beta.10",
 		"@appbaseio/reactivesearch": "^3.12.2",
+		"@loadable/component": "^5.15.2",
 		"@reach/router": "^1.2.1",
 		"emotion": "9.2.4",
 		"emotion-server": "9.2.4",