--- a/src/pages/index.js
+++ b/src/pages/index.js
@@ -15,7 +15,7 @@ const imagePrefix = 'https://opensource.appbase.io/reactivesearch/images/support
 
 const HomePage = ({ data, location }) => {
 	// Add meta title and description for this page here to overwrite the site meta data as set in the config
-
+// testing
 	return (
 		<>
 			<Layout