--- a/src/pages/index.js
+++ b/src/pages/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { graphql, navigate } from 'gatsby';
-
+import loadable from '@loadable/component'
 import PropTypes from 'prop-types';
 
-import TimelineOption from '@appbaseio/designkit/lib/molecules/TimelineOption';
+// import TimelineOption from '@appbaseio/designkit/lib/molecules/TimelineOption';
 import Grid from '@appbaseio/designkit/lib/atoms/Grid';
 import { Card } from '@appbaseio/designkit';
 import { Icon, Box } from '../components/common';
@@ -11,6 +11,7 @@ import { Layout } from '../components/common/layout';
 import { HomeHeader } from '../components/home';
 import { Spirit } from '../styles/spirit-styles';
 
+const TimelineOption = loadable(() => import('@appbaseio/designkit/lib/molecules/TimelineOption'))
 const imagePrefix = 'https://opensource.appbase.io/reactivesearch/images/support';
 
 const HomePage = ({ data, location }) => {