--- a/src/styles/custom.css
+++ b/src/styles/custom.css
@@ -1574,7 +1574,7 @@ to be custom HTML.
 .home-header-graphics .graphic {
 	padding: 15px;
 	padding-bottom: 7px;
-	width: 100%;
+	width: 190px;
 	border-radius: 4px;
 	background: #fff;
 	display: flex;