--- a/src/styles/custom.css
+++ b/src/styles/custom.css
@@ -1785,6 +1785,23 @@ to be custom HTML.
 	color: var(--darkgrey) !important;
 	font-size: 1.6rem !important;
 	font-weight: 600 !important;
+	font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell,
+		'Open Sans', 'Helvetica Neue', sans-serif;
+	padding: 0px 22px;
+	width: 185px;
+	height: 40px;
+	border-radius: 3px;
+	cursor: pointer;
+	user-select: none;
+	transition: all 0.3s ease 0s;
+	text-transform: uppercase;
+	letter-spacing: 0.01rem;
+	text-decoration: none;
+	white-space: nowrap;
+	background-color: #00f68e;
+	box-shadow: rgb(0 0 0 / 10%) 0px 3px 3px 0px;
+	border: none;
+	outline: none;
 }
 
 .video-btn {
@@ -1793,6 +1810,23 @@ to be custom HTML.
 	border: 0 !important;
 	font-weight: 600 !important;
 	background-color: #e4faff;
+	font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell,
+		'Open Sans', 'Helvetica Neue', sans-serif;
+	padding: 0px 22px;
+	width: 270px;
+	height: 40px;
+	border-radius: 3px;
+	cursor: pointer;
+	user-select: none;
+	transition: all 0.3s ease 0s;
+	text-transform: uppercase;
+	letter-spacing: 0.01rem;
+	text-decoration: none;
+	white-space: nowrap;
+	background-color: #00f68e;
+	box-shadow: rgb(0 0 0 / 10%) 0px 3px 3px 0px;
+	border: none;
+	outline: none;
 }
 
 .cursor-pointer {