… src/images/banner.webp
… static/images/app-cluster@3x.webp
… static/images/home-analytics@3x.webp
… static/images/home-search-relevancy@3x.webp
… static/images/swift-logo.webp
… yarn.lock