## Describe the change in this PR?
* https://www.notion.so/appbase/Docs-Performance-Issue-4724526ee1b8492bb7f659c0a57b2692
* PWA setup
* Chunks splitting
* Convert png to webp
* loadable components
* width and height for some images

Performance Improvement: 
Previous: mobile 30s, desktop 50s
Current: mobile 60s, desktop 60s